requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
pyTelegramBotAPI==3.7.3
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
